Move circuit breaker graph command
